refactor: glance folder
