+++ components/home/glance/Glance.tsx
